+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.4
 
+- Support examples/custom_draw_text.rb
 - Support stable `text` control nestable under `area`
 - Fix `color` property support for `string` to accept 255-based rgb values
 - Fix issue with alternating string colors in examples/basic_draw_text.rb